Move tags to common folder `Tag`
