+++ src/Html.php
@@ -32,9 +32,9 @@ use Yiisoft\Html\Tag\Input\Radio;
 use Yiisoft\Html\Tag\Label;
 use Yiisoft\Html\Tag\Li;
 use Yiisoft\Html\Tag\Link;
-use Yiisoft\Html\Tag\Media\Audio;
-use Yiisoft\Html\Tag\Media\Track;
-use Yiisoft\Html\Tag\Media\Video;
+use Yiisoft\Html\Tag\Audio;
+use Yiisoft\Html\Tag\Track;
+use Yiisoft\Html\Tag\Video;
 use Yiisoft\Html\Tag\Meta;
 use Yiisoft\Html\Tag\Noscript;
 use Yiisoft\Html\Tag\Ol;
